Fix copyright (finish), add CHANGES.md
The copyright information was finished (had authors but only one
listed). Not very important but it makes it more consistent and some of
the code is also based on code that I wrote (though that was in late
2024 early 2025), originally in the jparse repo (an artefact of how the
mkiocccentry toolkit was formed over time). Probably the comments should
have these details like I think libpr but doesn't.

Added CHANGES.md. Only since the most recent release (I did not update
the version - seemed no point).

I 'completed' the quote in the comments too.