improve this pull request
The `cpath` code was mostly written by us since about 1991 with some
more mods about 2008, 2014-2016, with some improvements in 2022-2025.
As such as will update the Copyright dates accordingly.

We only "dropped" in this code briefly to the
[mkiocccentry repo](https://github.com/ioccc-src/mkiocccentry)
while working in that repo.

We created the [cpath repo](https://github.com/lcn2/cpath) in part because
there were people in "stackoverflow" asking for such functionality,
because we had several places in our internal canonicalization was from
different code basis, and it was some to turn our private repo into a
public repo.

Similar to like `dbg` and `dyn_array`: `cpath` maybe used by this toolkit,
however it not part of this toolkit.  The only reason why those cloned
sub-directories reside in this repo is for the convenience of people
submitting to the contest.

Moreover, our deliberate intent for the Tolkien was just to provide the
fragment about wondering.  Nevertheless we will add a "_..._" to indicate
that the quote is only part of a sentence.

This commit will modify this PR, accordingly, as per the above.

Ran `make all depend tags picky check_man seqcexit test` to verify.